Refactor main into separate modules
