+++ src/mail.h
@@ -1,0 +1,10 @@
+#ifndef MAIL_H
+#define MAIL_H
+#include <time.h>
+void free_mail_list(void);
+void check_mail(void);
+/* helper functions exported for completeness */
+struct MailEntry;
+struct MailEntry *find_mail_entry(const char *path);
+void remember_mail_time(const char *path, time_t mtime);
+#endif /* MAIL_H */
