+++ src/repl.h
@@ -1,0 +1,5 @@
+#ifndef REPL_H
+#define REPL_H
+#include <stdio.h>
+void repl_loop(FILE *input);
+#endif /* REPL_H */
